api/requirements.txt
fastapi
uvicorn[standard]
llama-index
llama-index-vector-stores-postgres
llama-index-llms-groq
psycopg2-binary
